Update and move RecurringExecutionPlan and Test class
